Fix undefined behavior of applying an offset to a `NULL` pointer
The previous code assigned `0` to a pointer (`CovByPtr[0]`), which is the same as assigning `NULL`, generating a `NULL` pointer. Applying an offset to a `NULL` pointer, as is done in the cumulative loop where the `CovByPtr` values are collected, is undefined behavior. On most platforms, the `NULL` pointer will be converted back to a `0` when computing the offset, which is why this worked in the first place, but that behavior can't be relied on.